requirements.txt
django>=1.7
djangorestframework
